feat: added beans
